requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0          # async PostgreSQL driver
alembic==1.13.1

# Validation & settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Background tasks
celery==5.4.0
redis==5.0.4

# Financial calculations
python-dateutil==2.9.0
babel==2.15.0            # number/currency formatting

# PDF generation (statements, payoff letters)
weasyprint==62.1

# File handling & storage
boto3==1.34.102          # S3
python-magic==0.4.27

# HTTP client (for Investran API if needed)
httpx==0.27.0

# Utilities
orjson==3.10.3           # fast JSON serialisation
structlog==24.1.0        # structured logging
tenacity==8.3.0          # retry logic

# Dev / test
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0            # async test client
factory-boy==3.3.0
faker==25.0.0
ruff==0.4.4              # linting
mypy==1.10.0
